feat: encounters get all by tourist and tour

+++ encounters-api/model/social_encounter.go
@@ -12,7 +12,7 @@ type SocialEncounter struct {
 	ActiveTouristsIds *[]uint64 `json:"active_tourists_ids,omitempty" gorm:"type:bigint[]"`
 }
 
-func (se *SocialEncounter) CheckIfInRange(touristLongitude, touristLatitude float64, touristId int) int {
+func (se *SocialEncounter) CheckIfInRange(touristLongitude, touristLatitude float64, touristId uint64) int {
 	distance := math.Acos(math.Sin(math.Pi/180*(se.Encounter.Latitude))*math.Sin(math.Pi/180*touristLatitude)+math.Cos(math.Pi/180*se.Encounter.Latitude)*math.Cos(math.Pi/180*touristLatitude)*math.Cos(math.Pi/180*se.Encounter.Longitude-math.Pi/180*touristLongitude)) * 6371000
 	if distance > se.Range {
 		se.RemoveTourist(touristId)
@@ -23,13 +23,13 @@ func (se *SocialEncounter) CheckIfInRange(touristLongitude, touristLatitude floa
 	}
 }
 
-func (se *SocialEncounter) AddTourist(touristId int) {
+func (se *SocialEncounter) AddTourist(touristId uint64) {
 	if se.ActiveTouristsIds != nil && !contains(*se.ActiveTouristsIds, touristId) {
-		*se.ActiveTouristsIds = append(*se.ActiveTouristsIds, uint64(touristId))
+		*se.ActiveTouristsIds = append(*se.ActiveTouristsIds, touristId)
 	}
 }
 
-func (se *SocialEncounter) RemoveTourist(touristId int) {
+func (se *SocialEncounter) RemoveTourist(touristId uint64) {
 	if se.ActiveTouristsIds != nil && contains(*se.ActiveTouristsIds, touristId) {
 		index := indexOf(*se.ActiveTouristsIds, uint64(touristId))
 		*se.ActiveTouristsIds = append((*se.ActiveTouristsIds)[:index], (*se.ActiveTouristsIds)[index+1:]...)
@@ -48,7 +48,7 @@ func (se *SocialEncounter) ClearActiveTourists() {
 	*se.ActiveTouristsIds = []uint64{}
 }
 
-func contains(ids []uint64, id int) bool {
+func contains(ids []uint64, id uint64) bool {
 	for _, v := range ids {
 		if v == uint64(id) {
 			return true
